+++ MetaImage.go
@@ -42,13 +42,13 @@ func (meta *MetaImage) String() string {
 
 // ConvertToFile sends a request to the server and saves the resulting image in the given format.
 // Format can be one of the following: png, jpeg, gif, webp
-func (meta *MetaImage) ConvertToFile(format string, width int, height int, quality float64, fileName string) error {
+func (meta *MetaImage) ConvertToFile(format string, width int, height int, quality int, fileName string) error {
 	request := client.Get("http://127.0.0.1:" + ServerPort + "/")
 	request.Header("Content-Type", "image/"+meta.Format)
 	request.Header("Accept-Type", "image/"+format)
 	request.Header("Image-Width", fmt.Sprintf("%d", width))
 	request.Header("Image-Height", fmt.Sprintf("%d", height))
-	request.Header("Image-Quality", fmt.Sprintf("%f", quality))
+	request.Header("Image-Quality", fmt.Sprintf("%d", quality))
 	request.Body(meta.Data)
 	response, err := request.End()
